+++ Titania/Titania/Browser/X3DBrowserEditor.h
@@ -293,7 +293,7 @@ private:
 
 	///  @name Members
 
-	using UndoMatrixIndex = std::map <X3D::X3DTransformNodePtr, std::pair <X3D::Matrix4f, X3D::Vector3f>> ;
+	using UndoMatrixIndex = std::map <X3D::X3DTransformNodePtr, std::pair <X3D::Matrix4d, X3D::Vector3d>> ;
 
 	X3D::SFBool                        enabled;
 	std::unique_ptr <BrowserSelection> selection;
