+++ libtitania-x3d/Titania/X3D/Browser/Navigation/WalkViewer.h
@@ -106,8 +106,8 @@ private:
 	///  @name Operations
 
 	virtual
-	Vector3f
-	getTranslationOffset (const Vector3f &) final override;
+	Vector3d
+	getTranslationOffset (const Vector3d &) final override;
 
 	///  @name Static members
