+++ libtitania-x3d/Titania/X3D/Components/EnvironmentalSensor/ProximitySensor.h
@@ -173,7 +173,7 @@ private:
 	Fields fields;
 
 	X3DViewpointNode* viewpointNode;
-	Matrix4f          modelViewMatrix;
+	Matrix4d          modelViewMatrix;
 	bool              inside;
 
 };
